+++ R/godata.R
@@ -2,7 +2,9 @@
 ##'
 ##'
 ##' @title godata
-##' @param OrgDb OrgDb object
+##' @param OrgDb OrgDb object (will be removed in future, please use annoDb instead)
+##' @param annoDb GO annotation database, 
+##' can be OrgDb or a data.frame contains three columns of 'GENE', 'GO' and 'ONTOLOGY'.
 ##' @param keytype keytype
 ##' @param ont one of 'BP', 'MF', 'CC'
 ##' @param computeIC logical, whether computer IC
@@ -15,7 +17,7 @@
 ##' @importFrom methods new
 ##' @export
 ##' @author Guangchuang Yu
-godata <- function(OrgDb = NULL, keytype = "ENTREZID",
+godata <- function(OrgDb = NULL, annoDb = NULL, keytype = "ENTREZID",
                    ont, computeIC = TRUE,
                    processTCSS = FALSE, cutoff = NULL) {
     if (processTCSS) computeIC <- TRUE
@@ -23,23 +25,30 @@ godata <- function(OrgDb = NULL, keytype = "ENTREZID",
     ont <- toupper(ont)
     ont <- match.arg(ont, c("BP", "CC", "MF"))
 
-    if (is.null(OrgDb)) {
+    if (is.null(OrgDb) && is.null(annoDb)) {
         return(new("GOSemSimDATA",
         ont = ont
         ))
     }
 
-    OrgDb <- load_OrgDb(OrgDb)
-    kk <- keys(OrgDb, keytype = keytype)
-    message("preparing gene to GO mapping data...")
-    goAnno <- suppressMessages(
-        select(OrgDb,
-        keys = kk, keytype = keytype,
-        columns = c("GO", "ONTOLOGY")
-        )
-    )
+    if (!is.null(OrgDb)) {
+      warning("use 'annoDb' instead of 'OrgDb'")
+      annoDb <- OrgDb
+    }
+    if (is.character(annoDb)) {
+      annoDb <- load_OrgDb(annoDb) 
+    }
 
-    goAnno <- goAnno[!is.na(goAnno$GO), ]
+    md <- data.frame()
+    if (inherits(annoDb, 'OrgDb')) {
+      goAnno <- parse_orgDb(annoDb, keytype)
+      md <- metadata(annoDb)
+    } else if (inherits(annoDb, 'gson')) {
+      ## to be supported
+    } else { # for data.frame
+      goAnno <- check_goAnno(annoDb)
+    }
+
     goAnno <- goAnno[goAnno$ONTOLOGY == ont, ]
     if (computeIC) {
         message("preparing IC data...")
@@ -51,10 +60,10 @@ godata <- function(OrgDb = NULL, keytype = "ENTREZID",
     }
 
     res <- new("GOSemSimDATA",
-      keys = kk,
+      keys = unique(goAnno[,1]),
       ont = ont,
       geneAnno = goAnno,
-      metadata = metadata(OrgDb)
+      metadata = md
     )
     if (computeIC) {
         res@IC <- IC
@@ -64,6 +73,32 @@ godata <- function(OrgDb = NULL, keytype = "ENTREZID",
     }
 
     return(res)
+}
+
+check_goAnno <- function(goAnno) {
+  # check whether the data frame contains neccessary columns.
+
+  ## suppose 1st column is GENE ID and should contains GO and ONTOLOGY columns
+  ## maybe we should force names(goAnno)[1] == "GENE"
+  if (!all(c("GO", "ONTOLOGY") %in% names(goAnno))) {
+    stop("annoDb as a data.frame should contains 'GO' and 'ONTOLOGY' columns.")
+  }
+
+  return(goAnno)
+}
+
+parse_orgDb <- function(OrgDb, keytype) {
+    kk <- keys(OrgDb, keytype = keytype)
+    message("preparing gene to GO mapping data...")
+    goAnno <- suppressMessages(
+        select(OrgDb,
+        keys = kk, keytype = keytype,
+        columns = c("GO", "ONTOLOGY")
+        )
+    )
+
+    goAnno <- goAnno[!is.na(goAnno$GO), ]
+    return(goAnno)
 }
 
 ##' Class "GOSemSimDATA"